AI_modules/chat-module/requirements.txt
numpy==1.26.4
ninja==1.11.1.3
packaging==24.2
fastapi==0.111.0
Requests==2.32.3
uvicorn==0.30.1
pydantic==2.10.4
python-dotenv==1.0.1
openai==1.60.2
